+++ src/80-removeDuplicates.cpp
@@ -23,12 +23,12 @@ void test() {
     Solution s;
     vector<int> nums = {1, 1, 1, 2, 2, 3};
     vector<int> expected = {1, 1, 2, 2, 3};
-    assert(s.removeDuplicates(nums) == expected.size());
+    assert(s.removeDuplicates(nums) == static_cast<int>(expected.size()));
     assert(vector<int>(nums.begin(), nums.begin() + expected.size()) == expected);
 
     nums = {0, 0, 1, 1, 1, 1, 2, 3, 3};
     expected = {0, 0, 1, 1, 2, 3, 3};
-    assert(s.removeDuplicates(nums) == expected.size());
+    assert(s.removeDuplicates(nums) == static_cast<int>(expected.size()));
     assert(vector<int>(nums.begin(), nums.begin() + expected.size()) == expected);
 }
